moved stboot functions to pkg stboot removed log-calls and replaced with proper error handling.
